docker/requirements.txt
# Cython==3.0.4
async-timeout==4.0.2
dataclasses-json==0.5.8
fastjsonschema==2.18.1
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.1
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1
# ruamel-yaml-conda==0.15.80
# ruamel.yaml==0.17.35
# ruamel.yaml.clib==0.2.8
sentence-transformers==2.2.2
spacy-transformers==1.2.5
sqlparse==0.4.4
torch==2.0.1
torchdata==0.6.1
torchtext==0.15.2
torchvision==0.15.2
transformers==4.30.2
ujson==5.8.0
